Add new property and logic got overriddenByTaxonomy
modified the logic determining if a configured taxonomy overrides
a post_type configuration.

Where the post type is configured with default (ignored) values, and
an applicable taxonomy has configuration the the otherwise ignored
post_type's overriddenByTaxonomy setting was respected when set as
true, meaning that a taxonomy setting could be applied unexpectedly.

adding the overriddenByTaxonomy property to the class, and only setting
this value where the post_type has a non-default config ensures that
taxonomy settings aren't applied if the post_type otherwise has no config.