Change Action hook used
Wordpress 6.1 changes the execution order of the send_headers action
hook, moving it further back in the execution order, this means that
the functions that we are using to provide us with our context now return
the expected values.

Accordingly I am changing action hook used by the plugin so that we
are now using the appropriate and expected action hook instead of
using the template_redirect action hook.