Make the plugin less likely to override other plugins cache settings
This commit changes to plugin so that it checks the header array
set in the wp_headers filter, and if there is no logged in user and
the cache-control header has been set to no-cache then the setCacheHeader
method exits before it outputs any relvant headers.

Additionally, the setCacheHeader is now executed as the priority 1
task in the send_headers action loop, which will give any other plugins
which use the send_headers action, or any later action the opportunity
to override the cache-control header set by the plugin.

This should ensure that this plugin isn't overriding any other plugins
that are setting cache-control headers, providing that they are using
the wp_headers filter, the send_headers action, or an action that
executes after the send_headers action has been executed.

If a plugin is setting the cache-control header before the send_headers
action is executed then there is little we can do here, other than raise
a bug with the developer of that plugin.