(refactor) Add new Page class to hold all data on the current post/page
Currently, we're populating all the data about the post being viewed in
the `SendHeaders::pageProperties` property. This is populated by
`SendHeaders::getPageProperties()`.

It would be a step towards modularity in the plugin if we extracted that
data and put it in its own object, so I've added the `Page` class to do
that.

This commit does not implement the class, it just makes it available.

It's also worth noting that I've exactly replicated the logic as it
stands (e.g. for returning the post type), even where that logic appears
buggy. Instead, I've made a note in the comments of those cases, and
we'll fix them in separate commits.