(tests) Define a global $post object for the SendHeaders tests
Before: although `SendHeaders` references the global `$post` object, the
unit tests for the class never defined that object. They were still
passing because the cases tested relied on the properties of that object
being unpopulated.

Once we'd added tests for the new `Page` class, which defined different
global `$post` objects for different cases, those test `$post` objects
were then being re-used in the context of the `SendHeaders` tests,
because they weren't defining their own `$post` objects. The tests were
then failing.

By defining an empty global `$post` for all the `SendHeaders` tests,
they are now passing again. However, this is a strong indication that
the tests for that class are not providing full coverage of every
logic path.