(refactor) Implement `Page` object in `SendHeaders`
This refactor takes the `Page` object we'd created in previous commits
and implements it within the `SendHeaders` class. This is a first step
towards making this class more modular.

I've deliberately made the minimum possible changes to the tests for
`SendHeaders` in this commit, in order to demonstrate that this refactor
has not made any breaking changes - i.e. all I've done is pass in a mock
of the `Page` class (so the constructor doesn't throw an exception), and
change one test that mocks a callto  a method in `SendHeaders` that no
longer exists.

While it would be preferable to mock the calls to the `Page` instance
(rather than, as is happening now, mocking the calls that happen within
the `Page` instance's methods), these minimal changes confirm that the
logic within the new `Page` class precisely mirror the logic that was
previously within `SendHeaders`.