Rename Meta-CC headers x-debug-cache
It is conventional to use "X-" as a prefix for non-
standard HTTP headers, and this is quite common in
WordPress applications, including cache handling applications.

The use of X- as a prefix is generally not recommended
for new protocols, for reasons outlined here:

https://datatracker.ietf.org/doc/html/rfc6648

However, in this case we are not expecting the debug
headers in this plugin to become a widely adopted
convention, and using X- seems like the least surprising
naming to use.