(refactor) Implement `DeveloperMode` in `SendHeaders`
This change takes the `DeveloperMode` class we've created and uses it
within `SendHeaders` to create and output the developer mode-related
headers.

In this commit, I've deliberately not mocked the calls to
`DeveloperMode`, but rather left the existing mocks that now relate to
the calls that happen within `DeveloperMode`'s methods. This helps to
illustrate that while the implementation has changed, no behavioural
changes have occurred.

In a commit to follow, I'll mock the calls to `DeveloperMode` so that
we're limiting the focus of the unit tests to just the `SendHeaders`
class.